requirements.txt
requests>=2.28.1
setuptools
rich>=12.0.0
PyGithub
gitpython
ruamel.yaml
click>=7.1
dbt-core~=1.5.4
dbt-duckdb~=1.5.0
piperider[duckdb]@git+https://github.com/InfuseAI/piperider.git
